dashboard and navbar updates.
